+++ test/test.js
@@ -103,8 +103,120 @@ describe('parseurl(req)', function () {
   })
 })
 
-function createReq(url) {
+describe('parseurl.original(req)', function () {
+  it('should parse the request original URL', function () {
+    var req = createReq('/foo/bar', '/foo/bar')
+    var url = parseurl.original(req)
+    assert.equal(url.host, null)
+    assert.equal(url.hostname, null)
+    assert.equal(url.href, '/foo/bar')
+    assert.equal(url.pathname, '/foo/bar')
+    assert.equal(url.port, null)
+    assert.equal(url.query, null)
+    assert.equal(url.search, null)
+  })
+
+  it('should parse originalUrl when different', function () {
+    var req = createReq('/bar', '/foo/bar')
+    var url = parseurl.original(req)
+    assert.equal(url.host, null)
+    assert.equal(url.hostname, null)
+    assert.equal(url.href, '/foo/bar')
+    assert.equal(url.pathname, '/foo/bar')
+    assert.equal(url.port, null)
+    assert.equal(url.query, null)
+    assert.equal(url.search, null)
+  })
+
+  it('should parse req.url when originalUrl missing', function () {
+    var req = createReq('/foo/bar')
+    var url = parseurl.original(req)
+    assert.equal(url.host, null)
+    assert.equal(url.hostname, null)
+    assert.equal(url.href, '/foo/bar')
+    assert.equal(url.pathname, '/foo/bar')
+    assert.equal(url.port, null)
+    assert.equal(url.query, null)
+    assert.equal(url.search, null)
+  })
+
+  it('should return undefined missing req.url and originalUrl', function () {
+    var req = createReq()
+    var url = parseurl.original(req)
+    assert.strictEqual(url, undefined)
+  })
+
+  describe('when using the same request', function () {
+    it('should parse multiple times', function () {
+      var req = createReq('/foo/bar', '/foo/bar')
+      assert.equal(parseurl.original(req).pathname, '/foo/bar')
+      assert.equal(parseurl.original(req).pathname, '/foo/bar')
+      assert.equal(parseurl.original(req).pathname, '/foo/bar')
+    })
+
+    it('should reflect changes', function () {
+      var req = createReq('/foo/bar', '/foo/bar')
+      var url = parseurl.original(req)
+      var val = Math.random()
+
+      url._token = val
+      assert.equal(url._token, val)
+      assert.equal(url.pathname, '/foo/bar')
+
+      req.originalUrl = '/bar/baz'
+      url = parseurl.original(req)
+      assert.equal(url._token, undefined)
+      assert.equal(parseurl.original(req).pathname, '/bar/baz')
+    })
+
+    it('should cache parsing', function () {
+      var req = createReq('/foo/bar', '/foo/bar')
+      var url = parseurl.original(req)
+      var val = Math.random()
+
+      url._token = val
+      assert.equal(url._token, val)
+      assert.equal(url.pathname, '/foo/bar')
+
+      url = parseurl.original(req)
+      assert.equal(url._token, val)
+      assert.equal(url.pathname, '/foo/bar')
+    })
+
+    it('should cache parsing if req.url changes', function () {
+      var req = createReq('/foo/bar', '/foo/bar')
+      var url = parseurl.original(req)
+      var val = Math.random()
+
+      url._token = val
+      assert.equal(url._token, val)
+      assert.equal(url.pathname, '/foo/bar')
+
+      req.url = '/baz'
+      url = parseurl.original(req)
+      assert.equal(url._token, val)
+      assert.equal(url.pathname, '/foo/bar')
+    })
+
+    it('should cache parsing where href does not match', function () {
+      var req = createReq('/foo/bar ', '/foo/bar ')
+      var url = parseurl.original(req)
+      var val = Math.random()
+
+      url._token = val
+      assert.equal(url._token, val)
+      assert.equal(url.pathname, '/foo/bar')
+
+      url = parseurl.original(req)
+      assert.equal(url._token, val)
+      assert.equal(url.pathname, '/foo/bar')
+    })
+  })
+})
+
+function createReq(url, originalUrl) {
   return {
+    originalUrl: originalUrl,
     url: url
   };
 }